<Update>[PyutUIV2]: <Make sure AnchorPoints work>
[
This is bad mooky here. The Ogl objects were created withing having a Diagram
The legacy code deserialized the object while adding them to a frame. This
new code deserializes w/o reference to a frame
If we don't this the AnchorPoints are not on the diagram and lines ends are not
movable.
]

[https://github.com/hasii2011/PyUt/issues/426]